hypr-ai/requirements.txt
faiss-cpu
sentence-transformers
requests
numpy
tqdm
